requirements.txt
anytree==2.8.0
bitarray==1.2.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cycler==0.10.0
datasketch==1.5.0
decorator==4.4.1
elasticsearch==7.5.1
html5lib==1.0.1
idna==2.8
isodate==0.6.0
joblib==0.14.1
keepalive==0.5
kiwisolver==1.1.0
matplotlib==3.1.3
networkx==2.4
nltk==3.4.5
ontospy==1.9.8.3
pandas==1.0.0
PuLP==2.0
pyemd==0.5.1
pyfiglet==0.8.post1
pyparsing==2.4.6
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2019.3
PyYAML==5.4
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.22.0
scikit-learn==0.22.2.post1
six==1.13.0
snakecase==1.0.1
SPARQLWrapper==1.8.5
strsim==0.0.3
tqdm==4.42.0
urllib3==1.25.8
webencodings==0.5.1
# EmbDI
mlflow==1.9.0
gensim==3.8.3
text-unidecode==1.3
torchfile==0.1.0
faiss-cpu==1.6.3
scipy==1.7.1
numpy==1.21.2